+++ config/test.exs
@@ -11,6 +11,7 @@ config :portfolio_core, :manifest, manifest_path: "config/manifests/test.yml"
 config :portfolio_index,
   start_repo: false,
   start_boltx: false,
-  start_telemetry: false
+  start_telemetry: false,
+  env: :test
 
 config :logger, level: :warning